+++ src/main/java/com/example/package_delivery_system/data/repositories/PackageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PackageRepository extends JpaRepository<Office, Integer> {
+public interface PackageRepository extends JpaRepository<Office, Long> {
 }
